Делит JS-код на модули
