+++ index.html
@@ -231,6 +231,6 @@
     </div>
   </template>
 
-  <script src="main.js"></script>
+  <script src="js/main.js" type="module"></script>
 </body>
 </html>
